// libs/validator/src/lib/models/validation-type.enum.ts
export const enum ValidationTypeEnum {
  BAN_WORDS = 'banWords',
  APP_BAN_WORDS = 'appBanWords',
  APP_PASSWORD_SHOULD_MATCH = 'appPasswordShouldMatch',
  PASSWORD_SHOULD_MATCH = 'passwordShouldMatch',
  APP_UNIQUE_NICKNAME = 'appUniqueNickname',
  UNIQUE_NICKNAME = 'uniqueName',
  EMAIL = 'email',
  MINLENGTH = 'minlength',
  PATTERN = 'pattern',
  REQUIRED = 'required',
  REQUIRED_TRUE = 'requiredTrue',
}